Extract GIL-specific Boost.Test utilities from test/unit_test.hpp
Add separate header `test/unit_test_utility.hpp` for printers and
other utilities for better integration of GIL types with Boost.Test.
