+++ test/extension/numeric/convolve_cols.cpp
@@ -13,8 +13,9 @@
 
 #define BOOST_TEST_MODULE test_ext_numeric_colvolve_cols
 #include "unit_test.hpp"
-#include "core/image/test_fixture.hpp"
+#include "unit_test_utility.hpp"
 #include "test_fixture.hpp"
+#include "core/image/test_fixture.hpp"
 
 namespace gil = boost::gil;
 namespace fixture = boost::gil::test::fixture;
